1.2.10 craft slot moved to base class
