+++ src/main/java/mrriegel/storagenetwork/network/InsertMessage.java
@@ -46,7 +46,7 @@ public class InsertMessage implements IMessage, IMessageHandler<InsertMessage, I
         if (c instanceof ContainerRequest)
           t = w.getTileEntity(((ContainerRequest) c).tile.getMaster());
         else if (ctx.getServerHandler().player.openContainer instanceof ContainerRemote)
-          t = ((ContainerRemote) c).tileMaster;
+          t = ((ContainerRemote) c).getTileMaster();
         if (t instanceof TileMaster) {
           TileMaster tile = (TileMaster) t;
           int rest;
